+++ src/components/organisms/Accordion/AccordionHeader/AccordionHeader.test.tsx
@@ -1,9 +1,22 @@
 import React from 'react';
-import { render } from '@testing-library/react';
+import { fireEvent, render } from '@testing-library/react';
 import { axe } from 'jest-axe';
 import { Accordion } from '..';
 
 describe('<AccordionHeader />', () => {
+  it('fires onClick handler', () => {
+    const onClick = jest.fn();
+    const { getByTestId } = render(
+      <Accordion aria-label="test-accordion">
+        <Accordion.Header onClick={onClick} data-testid="header" id="one" index={0} textSize="base">
+          Header
+        </Accordion.Header>
+      </Accordion>,
+    );
+    fireEvent.click(getByTestId('header'));
+    expect(onClick).toHaveBeenCalledTimes(1);
+  });
+
   it('renders the component with no a11y violations', async () => {
     const { container, getByTestId } = render(
       <Accordion aria-label="test-accordion">