+++ src/index.ts
@@ -12,6 +12,7 @@ export { default as Fonts } from './components/styles/Fonts';
 export { default as GlobalStyles } from './components/styles/GlobalStyles';
 
 // Atoms
+export { default as Alert } from './components/atoms/Alert/Alert';
 export { default as AlertBox } from './components/atoms/AlertBox/AlertBox';
 export { default as Badge } from './components/atoms/Badge/Badge';
 export { default as Button, buttonStyle } from './components/atoms/Button/Button';
@@ -25,6 +26,7 @@ export { default as DropdownFiltered } from './components/molecules/DropdownFilt
 export { default as Spinner } from './components/atoms/Spinner/Spinner';
 export { default as ErrorMessage } from './components/atoms/ErrorMessage/ErrorMessage';
 export { default as InputLabel } from './components/atoms/InputLabel/InputLabel';
+export { default as InputRange } from './components/atoms/InputRange/InputRange';
 export { default as InputText } from './components/atoms/InputText/InputText';
 export { default as Text } from './components/atoms/Text/Text';
 export { default as Heading } from './components/atoms/Heading/Heading';