+++ src/styleguide-components/StyleGuide.js
@@ -1,10 +1,8 @@
 import React from 'react';
-import PropTypes from 'prop-types';
+import styled from 'styled-components';
 import { createGlobalStyle } from 'styled-components';
 import Logo from 'rsg-components/Logo';
 import Markdown from 'rsg-components/Markdown';
-import Styled from 'rsg-components/Styled';
-import cx from 'clsx';
 import Ribbon from 'rsg-components/Ribbon';
 import Version from 'rsg-components/Version';
 
@@ -14,58 +12,53 @@ import Fonts from '../components/styles/Fonts';
 const GlobalStyleGuideStyles = createGlobalStyle`
   /* System Fonts as used by GitHub */
   body {
-    font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
+    font-family: "Open Sans", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol";
   }
 `;
 
-const styles = ({ color, fontFamily, fontSize, sidebarWidth, mq, space, maxWidth }) => ({
-  root: {
-    backgroundColor: color.baseBackground,
+const sidebarWidth = 260;
+
+const Container = styled.div`
+  background-color: white;
+  padding-left: ${sidebarWidth + 20}px;
+  @media (max-width: 600px) {
+    padding-left: 0;
   },
-  hasSidebar: {
-    paddingLeft: sidebarWidth,
-    [mq.small]: {
-      paddingLeft: 0,
-    },
+`;
+const Main = styled.main`
+  max-width: 100%;
+  padding: 12px 16px;
+  margin: 0 auto;
+  display: block;
+`;
+
+const Sidebar = styled.div`
+  background-color: #1C2139;
+  border: 0;
+  position: fixed;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  width: ${sidebarWidth}px;
+  overflow: auto;
+  -webkit-overflow-scrolling: touch;
+  @media (max-width: 600px) {
+    position: static;
+    width: auto;
+    border-top-width: 1px;
+    padding-bottom: 0;
   },
-  content: {
-    maxWidth,
-    padding: [[space[2], space[4]]],
-    margin: [[0, 'auto']],
-    [mq.small]: {
-      padding: space[2],
-    },
-    display: 'block',
-  },
-  sidebar: {
-    backgroundColor: color.sidebarBackground,
-    border: [[color.border, 'solid']],
-    borderWidth: [[0, 1, 0, 0]],
-    position: 'fixed',
-    top: 0,
-    left: 0,
-    bottom: 0,
-    width: sidebarWidth,
-    overflow: 'auto',
-    '-webkit-overflow-scrolling': 'touch',
-    [mq.small]: {
-      position: 'static',
-      width: 'auto',
-      borderWidth: [[1, 0, 0, 0]],
-      paddingBottom: space[0],
-    },
-  },
-  logo: {
-    padding: space[2],
-    borderBottom: [[1, color.border, 'solid']],
-  },
-  footer: {
-    display: 'block',
-    color: color.light,
-    fontFamily: fontFamily.base,
-    fontSize: fontSize.small,
-  },
-});
+`;
+
+const LogoContainer = styled.div`
+  padding: 16px;
+`;
+
+const Footer = styled.footer`
+  display: block;
+  color: lightgray;
+  font-size: 12px;
+`;
 
 const footerText = homepageUrl => `
 <sub>Generated with [React Styleguidist](${homepageUrl}) for [Zopa's react-components](https://github.com/zopaUK/react-components). \
@@ -74,40 +67,28 @@ for reasonable and customary use in describing the origin of this library and re
 the [license](https://github.com/zopaUK/react-components/blob/master/LICENSE).</sub>
 `;
 
-export function StyleGuideRenderer({ classes, title, version, homepageUrl, children, toc, hasSidebar }) {
+export function StyleGuideRenderer({ title, version, homepageUrl, children, toc }) {
   return (
-    <div className={cx(classes.root, hasSidebar && classes.hasSidebar)}>
+    <Container>
       <GlobalStyles />
       <GlobalStyleGuideStyles />
       <Fonts />
-      <main className={classes.content}>
+      <Main>
         {children}
-        <footer className={classes.footer}>
+        <Footer>
           <Markdown text={footerText(homepageUrl)} />
-        </footer>
-      </main>
-      {hasSidebar && (
-        <div className={classes.sidebar}>
-          <div className={classes.logo}>
-            <Logo>{title}</Logo>
-            {version && <Version>{version}</Version>}
-          </div>
-          {toc}
-        </div>
-      )}
+        </Footer>
+      </Main>
+      <Sidebar>
+        <LogoContainer>
+          <Logo>{title}</Logo>
+          <Version>{version}</Version>
+        </LogoContainer>
+        {toc}
+      </Sidebar>
       <Ribbon />
-    </div>
+    </Container>
   );
 }
 
-StyleGuideRenderer.propTypes = {
-  classes: PropTypes.object.isRequired,
-  title: PropTypes.string.isRequired,
-  version: PropTypes.string,
-  homepageUrl: PropTypes.string.isRequired,
-  children: PropTypes.node.isRequired,
-  toc: PropTypes.node.isRequired,
-  hasSidebar: PropTypes.bool,
-};
-
-export default Styled(styles)(StyleGuideRenderer);
+export default StyleGuideRenderer;